+++ src/apps/nereus-finance/nereus-finance.module.ts
@@ -2,7 +2,6 @@ import { Register } from '~app-toolkit/decorators';
 import { AbstractApp } from '~app/app.dynamic-module';
 import { AaveV2AppModule } from '~apps/aave-v2/aave-v2.module';
 
-import { AvalancheNereusFinanceBalanceFetcher } from './avalanche/nereus-finance.balance-fetcher';
 import { AvalancheNereusFinancePositionPresenter } from './avalanche/nereus-finance.position-presenter';
 import { AvalancheNereusFinanceStableDebtTokenFetcher } from './avalanche/nereus-finance.stable-debt.token-fetcher';
 import { AvalancheNereusFinanceSupplyTokenFetcher } from './avalanche/nereus-finance.supply.token-fetcher';
@@ -14,7 +13,6 @@ import { NereusFinanceAppDefinition, NEREUS_FINANCE_DEFINITION } from './nereus-
   appId: NEREUS_FINANCE_DEFINITION.id,
   imports: [AaveV2AppModule],
   providers: [
-    AvalancheNereusFinanceBalanceFetcher,
     AvalancheNereusFinancePositionPresenter,
     AvalancheNereusFinanceStableDebtTokenFetcher,
     AvalancheNereusFinanceSupplyTokenFetcher,