+++ src/apps/plutus/arbitrum/plutus.lock.contract-position-fetcher.ts
@@ -53,9 +53,9 @@ const PLUTUS_LOCKS = [
 
 @Register.ContractPositionFetcher({ appId, groupId, network })
 export class ArbitrumPlutusLockContractPositionFetcher extends SingleStakingFarmTemplateContractPositionFetcher<PlutusLock> {
-  appId = PLUTUS_DEFINITION.id;
-  groupId = PLUTUS_DEFINITION.groups.lock.id;
-  network = Network.ARBITRUM_MAINNET;
+  appId = appId;
+  groupId = groupId;
+  network = network;
   groupLabel = 'Locked PLS';
 
   constructor(
@@ -108,7 +108,8 @@ export class ArbitrumPlutusLockContractPositionFetcher extends SingleStakingFarm
         const userPlsJonesShare = await multicall
           .wrap(rewardsContract)
           .calculateShare(address, epoch, rewardsForEpoch.plsJones);
-        if (Number(userPlsDpxShare) === 0 && Number(userPlsJonesShare) === 0) return [];
+        if (Number(userPlsDpxShare) === 0 && Number(userPlsJonesShare) === 0)
+          return [new BigNumber(0), new BigNumber(0)];
 
         const now = Date.now() / 1000;
         const vestedDuration =