+++ src/apps/spool/spool.definition.ts
@@ -7,7 +7,7 @@ export const SPOOL_DEFINITION = appDefinition({
   id: 'spool',
   name: 'Spool',
   description:
-    'Spool is decentralized middleware that allows anyone to create custom, diversified, and automated DeFi meta-strategies called “Spools”.',
+    'Spool is decentralized middleware that allows anyone to create custom, diversified, and automated DeFi meta-strategies called “Smart Vaults”.',
   url: 'https://app.spool.fi',
 
   groups: {
@@ -15,6 +15,11 @@ export const SPOOL_DEFINITION = appDefinition({
       id: 'vault',
       type: GroupType.POSITION,
       label: 'Spools',
+    },
+    staking: {
+      id: 'staking',
+      type: GroupType.POSITION,
+      label: 'Staking',
     },
   },