+++ src/apps/steak-hut/avalanche/steak-hut.pool.contract-position-fetcher.ts
@@ -3,7 +3,10 @@ import { BigNumberish } from 'ethers';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
-import { MasterChefTemplateContractPositionFetcher } from '~position/template/master-chef.template.contract-position-fetcher';
+import {
+  GetMasterChefTokenBalancesParams,
+  MasterChefTemplateContractPositionFetcher,
+} from '~position/template/master-chef.template.contract-position-fetcher';
 import { Network } from '~types/network.interface';
 
 import { SteakHutPool, SteakHutContractFactory } from '../contracts';
@@ -44,23 +47,31 @@ export class AvalancheSteakHutPoolContractPositionFetcher extends MasterChefTemp
     return contract.JOE();
   }
 
-  async getTotalAllocPoints(_contract: SteakHutPool): Promise<BigNumberish> {
+  async getTotalAllocPoints(): Promise<BigNumberish> {
     return 1;
   }
 
-  async getTotalRewardRate(_contract: SteakHutPool): Promise<BigNumberish> {
+  async getTotalRewardRate(): Promise<BigNumberish> {
     return 0;
   }
 
-  async getPoolAllocPoints(_contract: SteakHutPool, _poolIndex: number): Promise<BigNumberish> {
+  async getPoolAllocPoints(): Promise<BigNumberish> {
     return 0;
   }
 
-  async getStakedTokenBalance(address: string, contract: SteakHutPool, poolIndex: number): Promise<BigNumberish> {
-    return contract.userInfo(poolIndex, address).then(v => v.amount);
+  getStakedTokenBalance({
+    address,
+    contract,
+    contractPosition,
+  }: GetMasterChefTokenBalancesParams<SteakHutPool>): Promise<BigNumberish> {
+    return contract.userInfo(contractPosition.dataProps.poolIndex, address).then(v => v.amount);
   }
 
-  async getRewardTokenBalance(address: string, contract: SteakHutPool, poolIndex: number): Promise<BigNumberish> {
-    return contract.pendingJoe(poolIndex, address);
+  getRewardTokenBalance({
+    address,
+    contract,
+    contractPosition,
+  }: GetMasterChefTokenBalancesParams<SteakHutPool>): Promise<BigNumberish> {
+    return contract.pendingJoe(contractPosition.dataProps.poolIndex, address);
   }
 }