+++ src/position/position-fetcher.template-registry.ts
@@ -11,13 +11,13 @@ import { ContractPositionTemplatePositionFetcher } from './template/contract-pos
 @Injectable()
 export class PositionFetcherTemplateRegistry implements OnModuleInit {
   private appTokenTemplateRegistry = new Map<
-    Network,
-    Map<string, Map<string, AppTokenTemplatePositionFetcher<Erc20>>>
+    string,
+    Map<Network, Map<string, AppTokenTemplatePositionFetcher<Erc20>>>
   >();
 
   private contractPositionTemplateRegistry = new Map<
-    Network,
-    Map<string, Map<string, ContractPositionTemplatePositionFetcher<Erc20>>>
+    string,
+    Map<Network, Map<string, ContractPositionTemplatePositionFetcher<Erc20>>>
   >();
 
   constructor(@Inject(DiscoveryService) private readonly discoveryService: DiscoveryService) {}
@@ -25,32 +25,42 @@ export class PositionFetcherTemplateRegistry implements OnModuleInit {
   onModuleInit() {
     this.appTokenTemplateRegistry = buildTemplateRegistry(this.discoveryService, {
       template: AppTokenTemplatePositionFetcher,
-      keySelector: t => [t.network, t.appId, t.groupId] as const,
+      keySelector: t => [t.appId, t.network, t.groupId] as const,
     });
 
     this.contractPositionTemplateRegistry = buildTemplateRegistry(this.discoveryService, {
       template: ContractPositionTemplatePositionFetcher,
-      keySelector: t => [t.network, t.appId, t.groupId] as const,
+      keySelector: t => [t.appId, t.network, t.groupId] as const,
     });
   }
 
   getAppTokenTemplate({ network, appId, groupId }: { network: Network; appId: string; groupId: string }) {
-    return this.appTokenTemplateRegistry.get(network)?.get(appId)?.get(groupId) ?? null;
+    return this.appTokenTemplateRegistry.get(appId)?.get(network)?.get(groupId) ?? null;
   }
 
   getContractPositionTemplate({ network, appId, groupId }: { network: Network; appId: string; groupId: string }) {
-    return this.contractPositionTemplateRegistry.get(network)?.get(appId)?.get(groupId) ?? null;
+    return this.contractPositionTemplateRegistry.get(appId)?.get(network)?.get(groupId) ?? null;
   }
 
-  getTemplatesForApp(appId: string, network: Network) {
-    const appTokenTemplates = this.appTokenTemplateRegistry.get(network)?.get(appId);
-    const contractPositionTemplates = this.contractPositionTemplateRegistry.get(network)?.get(appId);
+  getTemplatesForApp(appId: string) {
+    const networks = Array.from(this.appTokenTemplateRegistry.get(appId)?.keys() ?? []);
+    return networks.flatMap(network => this.getTemplatesForAppOnNetwork(appId, network));
+  }
+
+  getAppHasTemplates(appId: string) {
+    const networks = Array.from(this.appTokenTemplateRegistry.get(appId)?.keys() ?? []);
+    return networks.some(network => this.getAppHasTemplatesOnNetwork(appId, network));
+  }
+
+  getTemplatesForAppOnNetwork(appId: string, network: Network) {
+    const appTokenTemplates = this.appTokenTemplateRegistry.get(appId)?.get(network);
+    const contractPositionTemplates = this.contractPositionTemplateRegistry.get(appId)?.get(network);
     return [...Array.from(appTokenTemplates?.values() ?? []), ...Array.from(contractPositionTemplates?.values() ?? [])];
   }
 
-  getAppHasTemplates(appId: string, network: Network) {
-    const appTokenTemplates = this.appTokenTemplateRegistry.get(network)?.get(appId);
-    const contractPositionTemplates = this.contractPositionTemplateRegistry.get(network)?.get(appId);
+  getAppHasTemplatesOnNetwork(appId: string, network: Network) {
+    const appTokenTemplates = this.appTokenTemplateRegistry.get(appId)?.get(network);
+    const contractPositionTemplates = this.contractPositionTemplateRegistry.get(appId)?.get(network);
     return !!appTokenTemplates || !!contractPositionTemplates;
   }
 }